FIX: Fixes dice_coef
