+++ unet3d/model.py
@@ -4,7 +4,7 @@ from keras.engine import Input, Model
 from keras.layers import Conv3D, MaxPooling3D, UpSampling3D, Activation
 from keras.optimizers import Adam
 
-from functools import partial
+from .metrics import dice_coef_loss, get_label_dice_coefficient_function, dice_coef
 
 K.set_image_data_format("channels_first")
 
@@ -18,6 +18,7 @@ def unet_model_3d(input_shape, pool_size=(2, 2, 2), n_labels=1, initial_learning
                   depth=4, n_base_filters=32, include_label_wise_dice_coefficients=False, metrics=(dice_coef)):
     """
     Builds the 3D UNet Keras model.
+    :param metrics: List metrics to be calculated during model training (default is dice coefficient).
     :param include_label_wise_dice_coefficients: If True and n_labels is greater than 1, model will report the dice
     coefficient for each label as metric.
     :param n_base_filters: The number of filters that the first layer in the convolution network will have. Following
@@ -73,27 +74,6 @@ def unet_model_3d(input_shape, pool_size=(2, 2, 2), n_labels=1, initial_learning
 
     model.compile(optimizer=Adam(lr=initial_learning_rate), loss=dice_coef_loss, metrics=metrics)
     return model
-
-
-def dice_coef(y_true, y_pred, smooth=1.):
-    y_true_f = K.flatten(y_true)
-    y_pred_f = K.flatten(y_pred)
-    intersection = K.sum(y_true_f * y_pred_f)
-    return (2. * intersection + smooth) / (K.sum(y_true_f) + K.sum(y_pred_f) + smooth)
-
-
-def dice_coef_loss(y_true, y_pred):
-    return -dice_coef(y_true, y_pred)
-
-
-def label_wise_dice_coefficient(y_true, y_pred, label_index):
-    return dice_coef(y_true[:, label_index], y_pred[:, label_index])
-
-
-def get_label_dice_coefficient_function(label_index):
-    f = partial(label_wise_dice_coefficient, label_index=label_index)
-    f.__setattr__('__name__', 'label_{0}_dice_coef'.format(label_index))
-    return f
 
 
 def compute_level_output_shape(n_filters, depth, pool_size, image_shape):
